validate: complete conversion of module to subpackage: split off code into bag.py, headbag.py
