+++ tests/multibag/validate/test_base.py
@@ -236,78 +236,6 @@ class TestValidator(test.TestCase):
         self.valid8r.ensure_valid()
 
 
-class TestBagValidator(test.TestCase):
-
-    def test_validate(self):
-        bagdir = os.path.join(datadir, "samplembag")
-        valid8r = val.BagValidator(bagdir)
-
-        results = valid8r.validate(val.ALL)
-        self.assertEqual(results.count_applied(), 1)
-        self.assertEqual(results.count_failed(), 0)
-        self.assertTrue(results.ok())
-
-        self.assertTrue(valid8r.is_valid())
-        valid8r.ensure_valid()
-
-    def test_validate_zip(self):
-        bagdir = os.path.join(datadir, "samplembag.zip")
-        valid8r = val.BagValidator(bagdir)
-
-        results = valid8r.validate(val.ALL)
-        self.assertEqual(results.count_applied(), 1)
-        self.assertEqual(results.count_failed(), 0)
-        self.assertTrue(results.ok())
-
-        self.assertTrue(valid8r.is_valid())
-        valid8r.ensure_valid()
-
-    def test_invalidate(self):
-        tempdir = tempfile.mkdtemp()
-        try:
-            bagdir = os.path.join(tempdir, "samplebag")
-            shutil.copytree(os.path.join(datadir, "samplembag"), bagdir)
-            os.rename(os.path.join(bagdir, "data"), os.path.join(bagdir, "goob"))
-            valid8r = val.BagValidator(bagdir)
-
-            results = valid8r.validate(val.ALL)
-            self.assertEqual(results.count_applied(), 1)
-            self.assertEqual(results.count_failed(), 1)
-            self.assertTrue(not results.ok())
-
-            self.assertTrue(not valid8r.is_valid())
-            with self.assertRaises(val.MultibagValidationError):
-                valid8r.ensure_valid()
-
-        finally:
-            shutil.rmtree(tempdir)
-
-class TestHeadBagValidator(test.TestCase):
-
-    def setUp(self):
-        self.tempdir = tempfile.mkdtemp()
-        self.bagdir = os.path.join(self.tempdir, "samplebag")
-        shutil.copytree(os.path.join(datadir, "samplembag"), self.bagdir)
-
-    def tearDown(self):
-        shutil.rmtree(self.tempdir)
-
-    def test_validate_version(self):
-        valid8r = val.HeadBagValidator(self.bagdir)
-        results = valid8r.validate_version()
-        self.assertEqual(results.count_applied(), 3)
-        self.assertTrue(results.ok())
-        
-        results = valid8r.validate_version(version="0.2")
-        self.assertEqual(results.count_applied(), 3)
-        self.assertTrue(not results.ok())
-        
-    def test_validate_reference(self):
-        valid8r = val.HeadBagValidator(self.bagdir)
-        results = valid8r.validate_reference()
-        self.assertEqual(results.count_applied(), 3)
-        self.assertTrue(results.ok())
-
